Idea for a Readline binding to increment/decrement a number
Want something similar to <C-a> and <C-x> in Vim. This function matches
and increments the number, but I'm currently stymied on how to insert
the number into the current line. There doesn't appear to be a way to do
that, except perhaps with the `shell-expand-line` function, and I don't
want to use that because I just want to insert the number, not expand
other parts of the shell line.

The other sticking point is how to read the value of the
`universal-argument`. There doesn't currently seem to be a way to do
that in a custom function.

I had a new idea for the number insertion problem, though. It's possible
I could replace the number, <C-c> the line, and reinsert a new copy with
the incremented/decremented number on the new shell line. That still
wouldn't resolve the `universal-argument` problem.