incdec: Handle point in the middle of subsequent numbers
Previously, we correctly incremented the first number if point was in
the middle of that number, but we did not do so for subsequent numbers.
This change adds similar handling for all numbers in the line.

Got a little lost here trying different loops and changes. I kind of
hacked it without heavily planning the algorithm ahead of time.

The idea is that if the point position is beyond the right side of the
match, we cut off the part of the line from the start to the end of the
last match, then try matching again. This allows us to pull out the
number around point so that we can increment it with the evaluated regex
that follows the loop.