incdec: Ideas for a loop based on match position
Looking into a different algorithm for incrementing matches that's
easier to use in both forward and backward contexts.

Got a few ideas to get a match nearest to the cursor position. We could
then take that value and increment it, and replace the original
substring with the new incremented value.