incdec: Move point if negative hyphen is added or removed
Previously, point would stay in the same position, causing it to look
like it moved when incrementing or decrementing between a negative
number and a positive number. This moves point by one to the right when
decrementing to a negative integer, or by one to the left when
incrementing to a positive integer. That makes the point seem like it
hasn't moved and making it look more natural.

For example, now point behaves like this:

    $ test 12 0
              ^
    _decrement_

    $ test 12 -1
               ^

    $ test 12 -1
               ^
    _increment_

    $ test 12 0
              ^

Before, it would do:

    $ test 12 0
              ^
    _decrement_

    $ test 12 -1
              ^

    $ test 12 -1
               ^
    _increment_

    $ test 12 0
               ^