100-increment-decrement: Want to return real start position from incdec
In order to set up point moving only when point is on or after the start
position of the changed number, I want the second return value from
`incdec()` to always give me the correct start position.

Currently, it returns 0 as the start position when backwards is on and
point is before the first number. It should return the start position of
the first number.